\name{wavelength2colour-package}
\alias{wavelength2colour-package}
\alias{wavelength2colour}
\docType{package}
\title{
\packageTitle{wavelength2colour}
}
\description{
\packageDescription{wavelength2colour}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wavelength2colour}
\packageIndices{wavelength2colour}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{wavelength2colour}

Maintainer: \packageMaintainer{wavelength2colour}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
