% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength2hex.R
\name{wavelength2hex}
\alias{wavelength2hex}
\title{Turn a wavelength (in nm) in to a hex colour value.}
\usage{
wavelength2hex(wavelength)
}
\arguments{
\item{wavelength}{a numerical wavelength to convert to hex.}
}
\value{
a hex representation of the wavelength.
}
\description{
Converts a wavelength to a hex value, via RGB.
}

