% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength2RGB.R
\name{wavelength2RGB}
\alias{wavelength2RGB}
\title{Turn a wavelength (in nm) in to an RGB vector.}
\usage{
wavelength2RGB(wavelength)
}
\arguments{
\item{wavelength}{a numerical wavelength to convert to RGB.}
}
\value{
a vector of three numeric values representing the RGB components.
}
\description{
Code adapted from \url{http://www.efg2.com/Lab/ScienceAndEngineering/Spectra.htm}
}

