\name{wavelength2colour-package}
\alias{wavelength2colour-package}
\alias{wavelength2colour}
\docType{package}
\title{
\packageTitle{wavelength2colour}
}
\description{
\packageDescription{wavelength2colour}
}
\details{
}
\author{
\packageAuthor{wavelength2colour}

Maintainer: \packageMaintainer{wavelength2colour}
}
\references{
}

\keyword{ package }
\seealso{
\code{\link[wavelength2colour]{wavelength2hex}}
\code{\link[wavelength2colour]{wavelength2RGB}}
}
\examples{
}
